+++ test/lib/absinthe/plug/document_provider/compiled_test.exs
@@ -3,24 +3,6 @@ defmodule Absinthe.Plug.DocumentProvider.CompiledTest do
   alias Absinthe.Plug.TestSchema
   alias Absinthe.Plug.DocumentProvider.Compiled
 
-  defmodule LiteralDocuments do
-    use Absinthe.Plug.DocumentProvider.Compiled
-
-    provide("1", """
-    query FooQuery($id: ID!) {
-      item(id: $id) {
-        name
-      }
-    }
-    """)
-
-    provide("2", """
-    query GetUser {
-      user
-    }
-    """)
-  end
-
   defmodule ExtractedDocuments do
     use Absinthe.Plug.DocumentProvider.Compiled
 
@@ -44,7 +26,10 @@ defmodule Absinthe.Plug.DocumentProvider.CompiledTest do
 
   test "works using documents provided as literals" do
     opts =
-      Absinthe.Plug.init(schema: TestSchema, document_providers: [__MODULE__.LiteralDocuments])
+      Absinthe.Plug.init(
+        schema: TestSchema,
+        document_providers: [Absinthe.Plug.TestLiteralDocuments]
+      )
 
     assert %{status: 200, resp_body: resp_body} =
              conn(:post, "/", %{"id" => "1", "variables" => %{"id" => "foo"}})
@@ -76,7 +61,10 @@ defmodule Absinthe.Plug.DocumentProvider.CompiledTest do
 
   test "context passed correctly to resolvers with documents provided as literals" do
     opts =
-      Absinthe.Plug.init(schema: TestSchema, document_providers: [__MODULE__.LiteralDocuments])
+      Absinthe.Plug.init(
+        schema: TestSchema,
+        document_providers: [Absinthe.Plug.TestLiteralDocuments]
+      )
 
     assert %{status: 200, resp_body: resp_body} =
              conn(:post, "/", %{"id" => "2"})
@@ -102,11 +90,13 @@ defmodule Absinthe.Plug.DocumentProvider.CompiledTest do
   end
 
   test ".get compiled" do
-    assert %Absinthe.Blueprint{} = Compiled.get(LiteralDocuments, "1")
-    assert %Absinthe.Blueprint{} = Compiled.get(LiteralDocuments, "1", :compiled)
+    assert %Absinthe.Blueprint{} = Compiled.get(Absinthe.Plug.TestLiteralDocuments, "1")
+
+    assert %Absinthe.Blueprint{} =
+             Compiled.get(Absinthe.Plug.TestLiteralDocuments, "1", :compiled)
   end
 
   test ".get source" do
-    assert @query == Compiled.get(LiteralDocuments, "1", :source)
+    assert @query == Compiled.get(Absinthe.Plug.TestLiteralDocuments, "1", :source)
   end
 end
